feat: Add light/dark mode toggle
This toggle using a view transition to transition between the two themes if supported. If not, the transitions is instant. A cookie is set so that the next time the user enters the site, the site is in their theme right away without having to re-click or wait for javascript to boot up.