feat: Add sidekiq with scraping job on all book show visits
When a user enters a show page for a book we start a job to scrape listings for that book, but only if it hasn't been done within the last 24 hours. This should keep books up to date as long as the books are visited relatively often. As an addition I will be adding a rake task to run scrapers for all books that haven't been updated in x days to catch books that don't get many visits.