fix(scraper): Remove stop words from boekennl slug
Boeken.nl removes certain stop words from their slugs. This seems to be a Drupal feature when using Pathauto. However, they seem to do this for all slugs but only for English stop words, even if the title is Dutch. The word "via" is one example that appears in both languages.