refactor: Move search over to books root page
To make it easier to search and filter on the same page, this removes the search controller entirely and adds search to the books filter. Search will use a query param rather than the nice url `zoeken/:query` it had before, but that will make filtering much easier. It will all happen on the same page.