fix: Prevent double binding of events on turbo load
Store all bound events in a WeakMap. Store elements with their ids. These ids are checked
before assigning the event listener, not applying the listener if it's already present.
A WeakMap is used so that elements are removed from the map when the element is removed.
This whole thing exists to prevent Turbo from re-binding events on turbo:load for elements
that already have the event attached to them.