fix: Don't set datetimes when calling scrape jobs
Previously we'd set the datetime right when calling the scrape. The idea was that this prevent more jobs from being called for the same object while the first job hadn't yet started.
The problem lies with jobs that never start to begin with, they'd be given a start date but never actually fine and thus become out of sync.

This moves update call to the jobs themselves but also adds an additional check. Worst case scenario the job gets called but immediately stops.