refactor: Use Solid Queue over Sidekiq
In order to simplify (and reduce costs), I'm switching over to Solid Queue. Annoyingly Solid Queue doesn't work on Windows, and getting it to work on WSL has been a struggle, so for the time being it's active storage on development, solid queue on prod. I'm sure that won't cause issues.