fix(filter): Allow array elements in filter
Previously the form data was converted to an object, but objects can't contain the same key multiple times, yet array elements in the url search param are exactly that. So instead they are now converted to an array and this array is looped over, placing valid items in a new array.