feat(scraper): Add max timeout to fetching listings
Listings that take a while to load are most likely not a valid page. Not necessarily, but this seems to be very true for boeken.nl, which takes more than 10 seconds to return whenever the page doesn't exist, where as regular requests are done in a few seconds. Setting a max timeout of 5 seconds will make those requests take a lot less long, as they are the bulk of the time spend right now.