refactor: Speed up several queries
Queries for books are slow because they need to include authors, genres, sources, listings, and more.
This speed them up significantly by adding a method that includes a load of always used data, as well as including subgenres several layers deep.

Additionally listings in cards are cached to a single variable before they are used.